Fix calls to Perl API when Perl is compiled with IMPLICIT_CONTEXT
In case Perl is built with PERL_IMPLICIT_CONTEXT, extensions that call
functions in the Perl API will need to pass the initial context argument
somehow. OpenBSD does not compile Perl this way, but the default package
on Ubuntu does.

This change fixes the calls introduced in ec4bd83 for Perl with
IMPLICIT_CONTEXT according to man(1) perlguts in 'the most efficient
way' by passing the context as an extra argument.