Second fix for calls to Perl API when Perl is compiled with IMPLICIT_CONTEXT
As 9eacf2988d3b856f35f4e91ebf1bf1d237a2ed88 did not work (see
0f24ab8d6265a379f52e28b7eaac1826c0ccd5b1), use the second way mentioned
in perlguts(1) to fetch the context in the functions. This is not as
fast as the third way but should be good enough for now.

This changes makes the XS module on Debian compilable again.