Enable github action
Use the github continuous integration framework to build the module and
run tests on push and pull requests. This first version uses one task to
install the requirements, build the software and test it. This could be
split up into different stages which pass the artifacts in the future,
if necessary.