Fix memory leak with failed client connect async calls.
The open62541 library invokes the async connect callback only if
the connection is successful.  Usually we malloc(3) the user data
in callback_async() and free(3) it in the open62541 callback after
we invoked the Perl callback.  But this free never happens when the
connection fails.

To fix this memory leak, extend the client object and additionally
save a pointer to the callback data structure.  When the client is
destroyed and if there is a callback structure, free it.  If the
callback is invoked, the callback data is still freed there, but
the pointer in the client is set to NULL.  So there will be no
doulbe free when the client is destroyed.  There is also the case
when callback_async() is called multiple times.  If there is an old
callback data at the client, it is replaced with the new one.