Free logger context from client and server destroy.
Extend the lifetime of the logger, it should log until the client
or server is destroyed.  For that the OPCUA_Open62541_Logger with
the Perl SVs is pointed to by the client or server config's logger
context.  The OPCUA_Open62541_Logger reference counts the client
or server config storage.  So the Perl logger keeps the Perl client
or server alive, the client or server destroy function is the last
one to be called.  If the Perl reference of the OPCUA_Open62541_Logger
is destroyed, the object is not freed.  This is is done later from
the client or server destroy function.