Do not generate more than one logger object.
Config getLogger() generates a OPCUA_Open62541_Logger object,
registers it in the client or server as logger context, and returns
a Perl reference to it.  If it was called a second time, the new
context replaced the old one, but was still referenced by its Perl
reference.  As a consequence, a following setCallback() used the
wrong context.  To avoid the Problem, getLogger() tries to use an
existing OPCUA_Open62541_Logger and returns it.  As we have distinct
Perl and C references, use an additional level of reference counting.