Embed logger struct into config.
To avoid complexity when freeing objects, embed the structures that
are the base for the Perl objects the same way as it is done in the
open62541 library.  The client or server structs are the only objects
that are malloced.  They contain the config which contains the
logger.  So the client or server must be freed last.  This is assured
by incrementing their Perl reference count each time a config or
logger is created.  When these are destroyed they decrement the
reference count and trigger client or server destruction.