Skip freeaddrinfo bug in client-connect-async test.
There is a bug in open62541 1.0.1 that crashes the client with a
segmentation fault.  It happens when the client delete() tries to
free an uninitialized addrinfo.  It is triggered by destroying a
client that never did a name lookup.  The OpenBSD port has a patch
that fixes the bug.  Use the buildinfo from the library to figure
out if we are affected.  Then skip the tests that trigger it.