Allocate UA_String and UA_ByteString data.
The memory of UA_String data should belong to open62541 instead of
Perl.  This allows to call UA_String_clear() or UA_String_delete()
later.  Also it will be possible to pass nested structs to open62541
and the library will free it.  To unpack UA_String and UA_ByteString
allocate memory and copy the data from SV.