Create mortal UA object in parameter typemap.
Move more of the parameter checks and memory management into the
typemap definitions.  The T_PTROBJ_PACKED typemap handles this.
The paramter name specifies whether it is a regular input, output
or optional output parameter.  The UA object for the input and
output parameter is generated on the fly and stored in a mortal SV.
The input parameter are unpacked automatically.  The mortal UA
object is available while the XS function is running.  After that
it is deleted by the DESTROY destructor.  The parameter checks and
explicit clear function in the funtions are no longer necessary.

Convert all UA_Node and UA_Variant parameter for now.