Test unicode conversion of UPC UA string and byte string.
Create unicode strings and utf-8 octets and pass them as strings
or byte strings to the test server.  Retrieve them with browse and
read request and check that they were not modified.  Write the
values to an UTF-8 text file and check that syswrite dies when
expected.  Convert the OPC UA hashes to YAML and test that it needs
the required encoding.