Implement client readDataTypeAttribute_async().
DataTypeAttribute is special as it has to convert a node to a number.
This has to be implemented manually.  Implement the same output
conversion logic as in UA_Client_readDataTypeAttribute() into
clientAsyncReadDataTypeCallback().  The outDataType reference check
in UA_Client_readDataTypeAttribute() has moved to the typemap.
Implement UA_Client_readDataTypeAttribute_async() like the generated
one, but call clientAsyncReadDataTypeCallback.