Pass node context reference to lifecycle constructor.
The open62541 API allows the constructor to create a node context.
Decouple the node context in server add...Node() by creating a SV
copy.  Then pass a reference to the Perl constructor.  Test that
the referenced variable can be changed and the new value is only
visible for the destructor.