Detect call to monitored item delete callback.
The open62541 library behaves inconsistently whether callbacks will
be called if registering the callback fails.  This make memory
managent hard, especially when multiple versions of the library
should be supported.
Register a pointer to the monitored item context and delete it if
the delete callback is called.  If the pointer is not NULL, the
callback data is still pending and must be freed.