Free callbacks if MonitoredItems_createDataChanges() failed.
New version of open62541 library behaves differently.  If registering
the data change callbacks fails, the delete callbacks are not called.
So they have to be deleted manually.  Add a test for this case.
Note that C memory is still leaking as the arrays of context and
callbacks are never freed by the final delete callback.