Use prolog and epilog for server add node functions.
There are many add node funtions in the server that share common
code.  Move this code into functions to make refactoring easier.
If there are neither constructor nor destructor, it is not necessary
to set up a node context SV.  Increment the references counter in
the prolog to ensure that it exists, even if the destructor has
deleted it.  Then the prolog can decrement the references counter,
taking into account if the destructort exists, has alread deleted
it or something went wrong during node creation.  The callback logic
of open26541 is so absurd that it is pure luck to find the correct
condition to destroy the SV.