Fix clientStateCallback client context.
Client getConfig() sets client context to the client SV passed in.
When the client variable goes out of scope, the sv becomes invalid.
Refresh the context every time the client connects, disconnects or
iterates to have an valid SV in clientStateCallback().  This fixes
the test case triggered by client-connect-async-refcount.t.