Implement pack and unpack with pointer arguments.
To construct the C structures in a sensible order, we have to
preallocate and link the pointers before we fill the values.  Perl
XS_pack and XS_unpack functions have a unsuitable API for that as
they pass objects by value and return structures.  This is not
optimal for programming in C.  Implement pack and unpack functions
that take pointers to UA and SV objects.  Generate XS_pack and
XS_unpack that call them.  The converters for the type tables use
void pointers and get the prefix table_pack_ and table_unpack_.