Use logging pointer from client or server config.
Open62541 version 1.4 uses a logging pointer instead of a logger
struct within the client and server config.  The logger clean
function is now called with the logger as parameter instead of the
context.  The Perl callback still uses the context argument.

This change allows OPCUA::Open62541 to compile with Open62541 1.4,
but the tests still fail.